:art: Implementa método base para escolha de opção em lista.

+++ Pinpad.Sdk.Test/NotAutomaticTests.cs
@@ -203,7 +203,7 @@ namespace Pinpad.Sdk.Test
             PinpadConnection conn = PinpadConnection.GetFirst();
             PinpadFacade facade = new PinpadFacade(conn);
 
-            int? value = facade.Keyboard.DataPicker.GetNumericValueInArray("Data Picker", 2, 4, 8, 16, 32, 64, 128);
+            short? value = facade.Keyboard.DataPicker.GetNumericValueInArray("Data Picker", 2, 4, 8, 16, 32, 64, 128);
 
             Debug.WriteLine(value.Value);
         }
